Add a read_count parameter to Arrays.
The `read_count` parameter is like `read_until` but supports the
circumstance where a user may want to read zero or more values.
`read_until`, unfortunately, does not allow the user to read 0 bytes
since it performs the read before the comparison and this results in at
least 1 byte read.