Fix tests for JRuby.
This fix includes two changes:

* Do not use fork in io_test.rb

The tests dependent on fork were just as easily implemented in a
single process using the pipe directly.

* Use a key other than "com" in lazy_test.rb

JRuby defines a top-level method "com" for our Java integration,
so this test's use of a :com key broke when method_missing did not
fire and our method was called. Changed name to :com1.