`to_binary_s` should always return ASCII-8BIT encoded strings
StringIO in Ruby <= 2.6.X didn't respect `Encoding.default_internal`, so
calling `read` on the StringIO wouldn't tag the string with the same
encoding.  Ruby 2.7 fixes this bug.  The problem is that `to_binary_s`
is returning a binary string and should always return an ASCII-8BIT
encoded string.

This patch changes the method to always return the ASCII-8BIT string
regardless of the value of `Encoding.default_internal`.

Here is the behavior of StringIO on Ruby 2.6:

```
$ ruby -v -rstringio -e'Encoding.default_internal = Encoding::UTF_8; io = StringIO.new("foo".force_encoding("ASCII-8BIT")); io.rewind; p io.read.encoding'
ruby 2.6.5p114 (2019-10-01 revision 67812) [x86_64-darwin19]
-e:1: warning: setting Encoding.default_internal
 #<Encoding:ASCII-8BIT>
```

Here it is on 2.7:

```
$ ruby -v -rstringio -e'Encoding.default_internal = Encoding::UTF_8; io = StringIO.new("foo".force_encoding("ASCII-8BIT")); io.rewind; p io.read.encoding'
ruby 2.7.0p0 (2019-12-25 revision 647ee6f091) [x86_64-darwin19]
-e:1: warning: setting Encoding.default_internal
  #<Encoding:UTF-8>
```

We can see 2.7 respects `default_internal`, but unfortunately that means
that `to_binary_s` would return the wrong encoding.